[tilemap] move canDesaturate check into generic utility
